Removed parent env delegation to fix crazy mem usage, now just copy all bindings (making this better later).

+++ src/pro_env_lookup.c
@@ -8,7 +8,7 @@
 
 struct pro_env_lookup
 {
-    unsigned int ref_count;
+    unsigned short ref_count;
     pro_env* value;
 };
